+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## (IN PROGRESS)
 
+* Limit selection of fiscal years during rollover to upcoming years. Refs UIF-286.
 * Allocation: set "To" budget as budget you are viewing by default. Refs UIF-284.
 * Allow only positive allocation to or from budget - Update field validation. Refs UIF-280.
 * Add label to transaction type "Rollover transfer". Refs UIF-282.